# src/main/java/com/skyhigh/checkin/dto/request/HoldSeatRequest.java
package com.skyhigh.checkin.dto.request;

import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class HoldSeatRequest {

    @NotNull(message = "Check-in ID is required")
    private UUID checkInId;
}