# src/main/java/com/skyhigh/checkin/dto/request/LoginRequest.java
package com.skyhigh.checkin.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class LoginRequest {

    @NotBlank(message = "Booking reference is required")
    @Pattern(regexp = "^[A-Z0-9]{6}$", message = "Booking reference must be 6 alphanumeric characters")
    private String bookingReference;

    @NotBlank(message = "Last name is required")
    private String lastName;

    @NotBlank(message = "Email is required")
    @Email(message = "Invalid email format")
    private String email;
}