# src/main/java/com/skyhigh/checkin/dto/request/StartCheckInRequest.java
package com.skyhigh.checkin.dto.request;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class StartCheckInRequest {

    @NotNull(message = "Flight ID is required")
    private UUID flightId;

    @NotBlank(message = "Booking reference is required")
    private String bookingReference;
}